QRGeneratePrintKit/requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
idna==3.10
lxml==6.0.0
numpy==2.3.1
panda==0.3.1
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
qrcode==8.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
use==0.1.0
